+++ dluhc-component-library/src/components/formComponents/types.ts
@@ -1,0 +1,4 @@
+export type RadioOption<T extends string | boolean> = {
+  label: string;
+  value: T;
+};
