+++ dluhc-component-library/src/components/siteSelectionForm/types.ts
@@ -46,8 +46,3 @@ export type Widget = "map" | "radio";
 export type UiPropertySchema = { "ui:widget"?: Widget };
 
 export type UiSchema = Record<string, UiPropertySchema>;
-
-export type RadioOption<T extends string | boolean> = {
-  label: string;
-  value: T;
-};
